Suggestion to introduce CIDs
The aim of this PR is to assert that CIDs are points of extension to the Ethernet frame transport, rather than some generalised concept for arbitrary PDUs - while at the same time allowing flexibility in how these are eventually used. The chief goal is to avoid the question "what else can be can carried instead of an Ethernet PDU?".