Proposal to rework VLAN Text
This proposal rewrites the VLAN text as an example of how the processing is performed on either side of the proxy. 
In this, I used "MAY" for strip the VLAN, however it could indeed be "RECOMMENDED" to strip the VLAN, but then I wondered if that asked for more guidance on how to do that consistently (which would also be a fine addition to this text).